+++ src/freetds/src/tds/config.c
@@ -582,7 +582,7 @@ tds_parse_conf_section(const char *option, const char *value, void *param)
 		long flags;
 		flags = strtol(value, &end, 0);
 		if (*value != '\0' && *end == '\0' && flags != LONG_MIN && flags != LONG_MAX)
-			connection->debug_flags = flags;
+			connection->debug_flags = (int)flags;
 	} else if (!strcmp(option, TDS_STR_TIMEOUT) || !strcmp(option, TDS_STR_QUERY_TIMEOUT)) {
 		if (atoi(value))
 			connection->query_timeout = atoi(value);
